// client/src/utils/types.ts
export type UserPeer = {
    id: string;
}



export type User = {
    id: number;
    username: string;
    firstName: string;
    lastName: string;
    profile?: Profile;
    presence?: UserPresence;
    peer: UserPeer;
};

export type UserPresence = {
    id: number;
    statusMessage?: string;
    showOffline: boolean;
}

export type Conversation = {
    id: number;
    creator: User;
    recipient: User;
    createdAt: string;
    lastMessageSent: MessageType;
};

export type CreateConversationParams = {
    username: string;
    message: string;
};

export type MessageType = {
    id: number;
    content: string;
    createdAt: string;
    author: User;
    conversation: Conversation;
};
// ...
export type Group = {
    id: number;
    title?: String;
    users: User[];
    creator: User;
    owner: User;
    messages: GroupMessageType[];
    createdAt: number;
    lastMessageSent: MessageType;
    lastMessageSentAt: Date;
}

export type GroupMessageType = {
    id: number;
    content: string;
    createdAt: string;
    author: User;
    group: Group;


}
// ...
export type UserContextMenuActionType = 'kick' | 'transfer_ower' | 'profile';
// ...
export type Friend = {
    id: number;
    sender: User;
    receiver: User;
    createdAt: number;
}
// ...
export type UserSidebarRouteType = 'conversations' | 'friends' | 'connections' | 'settings' | 'calls';

export type UserSidebarItemType = {
    id: UserSidebarRouteType;
    pathname: string;

}

export type SettingsSidebarRouteType = | 'profile' | 'security' | 'notifications' | 'integrations' | 'appearance';
// ...
export type Profile = {
    id: number;
    about?: string;
    avatar?: string;
    banner?: string;
}